src/search-forecast/requirements.txt
grpcio
grpcio-tools
requests
circuitbreaker